Mise à jour depuis phpbot

+++ irc/modules/ModChannelUsers.php
@@ -1,23 +1,31 @@
 <?php
  class ModChannelUsers extends Module {
-  private $bInWho = false;
+  private $abInWho = array();
   private $aWhoBuffer = array();
   private $oInstance;
-  private $aiOldUsers = array();
+  
+  protected function _emptyChannel($sChannel) {
+   $oChannelUsers = new dbChannelUsers();
+   $aoChannelUsers = $oChannelUsers->writable()->select()->where('channel = ?', $sChannel)->fetchAll();
+   foreach ($aoChannelUsers as $oChannelUser) {
+    $oChannelUser->delete();
+   }
+  }
   
   public function onLoad() {
    $this->oInstance = dbInstance::get('site');
+   $this->oInstance->exec('TRUNCATE TABLE `channel_users`;');
   }
 
   public function onWhoLine(ParsedMask $oWho, $sTarget, $sFlags, $sDescription){
-   if ($this->bInWho) {
-    $this->aWhoBuffer[] = array($oWho, $sTarget);
+   if (!empty($this->abInWho[$sTarget])) {
+    $this->aWhoBuffer[$sTarget][] = array($oWho);
    }
   }
   
   public function onJoin(ParsedMask $oWho, $sChannel){
    $oChannelUsers = new dbChannelUsers();
-   
+   var_dump(func_get_args());
    $oChannelUser = $oChannelUsers->writable()->select('*, IF(date_autologin IS NULL, 0, DATE_ADD(date_autologin, INTERVAL 10 MINUTE) > NOW()) AS autologin')->where('channel = ? AND nick = ?', $sChannel, $oWho->getNick())->fetch();
    if (!$oChannelUser) {
     $oChannelUser = $oChannelUsers->create();
@@ -36,40 +44,52 @@
    $oChannelUser->save();
    
    if ($this->isMe($oWho->getNick())) {
-    $this->bInWho = true;
+    $this->abInWho[$sChannel] = true;
     $this->who($sChannel);
    }
   }
   
   public function onPart(ParsedMask $oWho, $sChannel, $sMessage){
-   $oChannelUsers = new dbChannelUsers();
-   if ($oChannelUser = $oChannelUsers->writable()->select()->where('channel = ? AND nick = ?', $sChannel, $oWho->getNick())->fetch()) {
-    $this->aiOldUsers[(string)$oWho] = (int)$oChannelUser->id_irpg_user;
-    $oChannelUser->delete();
+   if ($this->isMe($oWho->getNick())) {
+    $this->_emptyChannel($sChannel);
+   } else {
+    $oChannelUsers = new dbChannelUsers();
+    if ($oChannelUser = $oChannelUsers->writable()->select()->where('channel = ? AND nick = ?', $sChannel, $oWho->getNick())->fetch()) {
+     $this->aiOldUsers[(string)$oWho] = (int)$oChannelUser->id_irpg_user;
+     $oChannelUser->delete();
+    }
    }
   }
   
   public function onKick(ParsedMask $oWho, $sChannel, $sToNick, $sMessage){
-   $oChannelUsers = new dbChannelUsers();
-   if ($oChannelUser = $oChannelUsers->writable()->select()->where('channel = ? AND nick = ?', $sChannel, $oWho->getNick())->fetch()) {
-    $this->aiOldUsers[(string)$oWho] = (int)$oChannelUser->id_irpg_user;
-    $oChannelUser->delete();
+   if ($this->isMe($oWho->getNick())) {
+    $this->_emptyChannel($sChannel);
+   } else {
+    $oChannelUsers = new dbChannelUsers();
+    if ($oChannelUser = $oChannelUsers->writable()->select()->where('channel = ? AND nick = ?', $sChannel, $oWho->getNick())->fetch()) {
+     $this->aiOldUsers[(string)$oWho] = (int)$oChannelUser->id_irpg_user;
+     $oChannelUser->delete();
+    }
    }
   }
   
   public function onQuit(ParsedMask $oWho, $sMessage){
-   $oChannelUsers = new dbChannelUsers();
-   if ($aoChannelUsers = $oChannelUsers->writable()->select()->where('nick = ?', $oWho->getNick())->fetchAll()) {
-    foreach ($aoChannelUsers as $oChannelUser) {
-     if ($this->isGameChannel($oChannelUser->channel)) {
-      if ($sMessage == '*.net *.split' || $sMessage == 'registered') {
-       $oChannelUser->date_autologin = new dbDontEscapeString('NOW()');
-       $oChannelUser->save();
-       continue;
+   if ($this->isMe($oWho->getNick())) {
+    $this->oInstance->exec('TRUNCATE TABLE `channel_users`;');
+   } else {
+    $oChannelUsers = new dbChannelUsers();
+    if ($aoChannelUsers = $oChannelUsers->writable()->select()->where('nick = ?', $oWho->getNick())->fetchAll()) {
+     foreach ($aoChannelUsers as $oChannelUser) {
+      if ($this->isGameChannel($oChannelUser->channel)) {
+       if ($sMessage == '*.net *.split' || $sMessage == 'registered') {
+        $oChannelUser->date_autologin = new dbDontEscapeString('NOW()');
+        $oChannelUser->save();
+        continue;
+       }
       }
+      $this->aiOldUsers[(string)$oWho] = (int)$oChannelUser->id_irpg_user;
+      $oChannelUser->delete();
      }
-     $this->aiOldUsers[(string)$oWho] = (int)$oChannelUser->id_irpg_user;
-     $oChannelUser->delete();
     }
    }
   }
@@ -77,14 +97,14 @@
   public function onNick(ParsedMask $oWho, $sNewNick){
    $oChannelUsers = new dbChannelUsers();
    $aoChannelUsers = $oChannelUsers->writable()->select()->where('nick = ?', $oWho->getNick())->fetchAll();
-   foreach ($aoChannelUser as $oChannelUser) {
+   foreach ($aoChannelUsers as $oChannelUser) {
     $oChannelUser->nick = $sNewNick;
     $oChannelUser->save();
    }
   }
 
   public function onEndOfWho($sTarget){ // Normalement c'est exécute qu'une fois lorsque le bot join le salon
-   if ($this->bInWho) {
+   if (!empty($this->aWhoBuffer[$sTarget])) { // Si une fin de who est en cours sur ce salon
     $oChannelUsers = new dbChannelUsers();
     
     $aWasLoggedUsers = array();
@@ -95,38 +115,35 @@
      }
     }
 
-    $this->oInstance->exec('TRUNCATE TABLE channel_users;');
-
-    foreach ($this->aWhoBuffer as $aWhoUser) {
-     if ($sTarget == $aWhoUser[1]) { // On sait jamais
-      $oWho = $aWhoUser[0];
-
+    foreach ($this->aWhoBuffer[$sTarget] as $aWhoUser) {
+     $oWho = $aWhoUser[0];
+     
+     $oChannelUser = $oChannelUsers->writable()->select()->where('channel = ? AND nick = ?', $sTarget, $oWho->getNick())->fetch();
+     if (!$oChannelUser) {
       $oChannelUser = $oChannelUsers->create();
-      
       $oChannelUser->channel = $sTarget;
       $oChannelUser->nick = $oWho->getNick();
-      $oChannelUser->user = $oWho->getUser();
-      $oChannelUser->host = $oWho->getHost();
-      
-      if (isset($aWasLoggedUsers[(string)$oWho])) { // Relog automatique
-       $oChannelUser->id_irpg_user = $aWasLoggedUsers[(string)$oWho];
-       $oChannelUser->date_autologin = new dbDontEscapeString('NULL');
-       $oChannelUser->save();
-       
-       ModuleManager::dispatch('doUserLogin', $oWho, $aWasLoggedUsers[(string)$oWho], true);
-      } else {
-       $oChannelUser->id_irpg_user = new dbDontEscapeString('NULL');
-       $oChannelUser->date_autologin = new dbDontEscapeString('NULL');
-       $oChannelUser->save();
-      }
-
      }
+     $oChannelUser->user = $oWho->getUser();
+     $oChannelUser->host = $oWho->getHost();
+     
+     if (isset($aWasLoggedUsers[(string)$oWho])) { // Relog automatique
+      $oChannelUser->id_irpg_user = $aWasLoggedUsers[(string)$oWho];
+      $oChannelUser->date_autologin = new dbDontEscapeString('NULL');
+      $oChannelUser->save();
+      
+      ModuleManager::dispatch('doUserLogin', $oWho, $aWasLoggedUsers[(string)$oWho], true);
+     } else {
+      $oChannelUser->id_irpg_user = new dbDontEscapeString('NULL');
+      $oChannelUser->date_autologin = new dbDontEscapeString('NULL');
+      $oChannelUser->save();
+     }
+
     }
+    
+    unset($this->aWhoBuffer[$sTarget]);
+    unset($this->abInWho[$sTarget]);
     // A ce point, liste des utilisateurs présents sur le salon est à jour
-    
-    $this->aWhoBuffer = array();
-    $this->bInWho = false;
-    
     ModuleManager::dispatch('onUserListUpdated');
    }
   }
@@ -144,5 +161,12 @@
   public function onCtcpReply(ParsedMask $oWho, $sTarget, $sMessage){}
   public function onNamesLine($sChannel, array $aUsers){}
   public function onRaw($iRaw, $sArguments){}
+  public function onUnload(){}
+  
+  public function getMaskFromNick($sNick) {
+   $oChannelUsers = new dbChannelUsers();
+   if ($oChannelUser = $oChannelUsers->select('CONCAT(nick, "!", user, "@", host) AS mask')->where('nick = ?', $sNick)->fetch()) {
+    return new ParsedMask((string)$oChannelUser->mask);
+   }
+  }
  }
-?>
